+++ src/public_health_framework/__init__.py
@@ -1,6 +1,6 @@
 """Public Health Framework."""
 
-__version__ = "0.7.0a1"
+__version__ = "0.7.0a2"
 
 from .application import PHFrame
 from .config import DatasetSchema, FieldSchema, ProjectConfig